refactor(core): move BinaryBodySchema + binary schemas to core/schemas
These schemas describe `application/octet-stream` request/response bodies
and are completely cloud-agnostic — every distilled SDK that needs raw
binary IO (S3, GCS, R2, Azure Blob, …) wants the same union. Hosting them
in `@distilled.cloud/core/schemas` lets every package share one
`BinaryBodySchema` / `BinaryStreamResponseSchema` contract.

`@distilled.cloud/cloudflare`'s `schemas` module now re-exports the
binary primitives from core and only owns the cloudflare-specific
multipart upload schemas (`FileSchema`, `UploadableSchema`).
